fix(media): address review feedback on overrideMicConstraints
Three changes per @IsmaelMartinez's review:

- Swap module registration order in preload.js so overrideMicConstraints
  registers before disableAutogain. With the prior order, disableAutogain
  re-asserted autoGainControl=false after overrideMicConstraints had
  written the user override, defeating the documented precedence rule.

- Extract setLegacyChromeConstraint / setConstraint into a shared
  _micConstraintHelpers.js module. Both files imported the same bodies
  (setConstraint byte-identical, setLegacyChromeConstraint differed only
  by comment wording before extraction, kept disableAutogain's wording).
  Resolves the SonarQube duplication finding.

- Replace the misleading "Don't mutate the original reference" comment
  on the applyConstraints wrapper. The wrapper aliases the same
  constraint object and overrideConstraints does mutate it; new wording
  documents the intent honestly.

No behavior change to the no-arg applyConstraints path: Teams does not
call applyConstraints without args, so the && constraints guard stays
as-is per maintainer guidance.
