+++ app/browser/preload.js
@@ -347,8 +347,8 @@ document.addEventListener('DOMContentLoaded', async () => {
       { name: "timestampCopyOverride", path: "./tools/timestampCopyOverride" },
       { name: "trayIconRenderer", path: "./tools/trayIconRenderer" },
       { name: "mqttStatusMonitor", path: "./tools/mqttStatusMonitor" },
-      { name: "disableAutogain", path: "./tools/disableAutogain" },
       { name: "overrideMicConstraints", path: "./tools/overrideMicConstraints" },
+      { name: "disableAutogain", path: "./tools/disableAutogain" },
       { name: "speakingIndicator", path: "./tools/speakingIndicator" },
       { name: "cameraResolution", path: "./tools/cameraResolution" },
       { name: "cameraAspectRatio", path: "./tools/cameraAspectRatio" },
